refactor(validate): utils로 파일 이름 변경
validate로 나누려니 기준이 모호한 상수들이 있어서 파일 이름을 utils로 변경했습니다.
