+++ src/constants/utils.js
@@ -4,3 +4,4 @@ export const VALID_LOTTERY_NUM = 6;
 export const VALID_LOWEST_NUM = 1;
 export const VALID_HIGHEST_NUM = 45;
 export const MAX_LIMIT = 1000000;
+export const PREFIX_ERRMSG = '[ERROR]';
